Use old approach to convert bytes to strings
