+++ .github/workflows/test.yml
@@ -17,7 +17,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-go@v3
         with:
-          go-version: '1.20'
+          go-version: 1.19
       - uses: golangci/golangci-lint-action@v3
         with:
           version: v1.50
